move html filter and validator
